+++ gmapping_glass/gmapping_export/scanmatcher/scanmatcher.h
@@ -28,7 +28,7 @@ class ScanMatcher{
 			(unsigned int beams, double* angles, const OrientedPoint& lpose);
 		void setMatchingParameters
 			(double urange, double range, double sigma, int kernsize, double lopt, double aopt, int iterations, double likelihoodSigma=1, unsigned int likelihoodSkip=0 );
-    void setGlassDetectionParameters(double trigInt, double intDelta, double profWidth);
+    void setGlassDetectionParameters(double trigInt, double intDelta, int profWidth);
 		void invalidateActiveArea();
 		void computeActiveArea(ScanMatcherMap& map, const OrientedPoint& p, const double* readings);
